test(evm): move handler tests to separate file
